+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Moved Item Registration to DeferredRegistry
 - Updated block and block entity registration
 - Updated ChickenTeachHandler
+- Updated LiquidEggFluidWrapper to FluidEggWrapper
 
 ### Deprecated